+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 <<<<<<< HEAD
+<<<<<<< HEAD
 - New `*-LabSnippet` cmdlets!
 - New PreInstallationActivity parameter for Add-LabMachineDefinition
   - Immediately before a role is deployed, the PreInstallationActivity is executed
@@ -22,6 +23,9 @@
 - Replaced dynamic parameters with user-friendly TEPP Argument Completers.
 - Update Exchange2019 custom role to use public download.
 
+=======
+- Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
+>>>>>>> 6eaed7b1 (Added SQL Server 2019 support for Azure DevOps Server)
 ### Fixes
 - Fixed #1087. AL now supports WIM and ESD files in Windows ISOs.
 - Fixed #1115, the validator for AzDevOps/TFS now also validates SQLServer2019