+++ CHANGELOG.md
@@ -9,8 +9,10 @@
   - Machines without roles will get their PreInstallActivity executed after all domains are deployed
   - For both pre and post, Get-LabInstallationActivity can be used, and can even include custom roles
   - Thanks to the awesomeness that is @friedrichweinmann and his PSFramework, we now have Variables and Functions that can be passed to Get-LabInstallationActivity
+- Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
 
 ### Fixes
+- Fixed #1033, NonExistingDnsServerAssigned is now only called if domain controllers are defined in the lab.
 
 ## 5.34.0 (2021-03-30)