+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
 - Added product keys for 'Windows 10 Education' and 'Windows 10 Pro Education'.
 - Setting lab administrative accounts to never expire
+- MAC addresses are now configurable
 
 ### Fixes
 - Fixed #1033, NonExistingDnsServerAssigned is now only called if domain controllers are defined in the lab.