+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
 - Added product keys for 'Windows 10 Education' and 'Windows 10 Pro Education'.
 - MAC addresses are now configurable
-- A choice selection of WSMAN settings can now be configuration
+- A choice selection of WSMAN settings can now be configured
   - To see which: Get-PSFConfig -FullName AutomatedLab.WinRm*
 
 ### Fixes