+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Thanks to the awesomeness that is @friedrichweinmann and his PSFramework, we now have Variables and Functions that can be passed to Get-LabInstallationActivity
 - Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
 - Added product keys for 'Windows 10 Education' and 'Windows 10 Pro Education'.
+- Setting lab administrative accounts to never expire
 - MAC addresses are now configurable
 
 ### Fixes