+++ CHANGELOG.md
@@ -6,6 +6,28 @@
 
 ### Fixes
 
+## 5.35.0 (2021-04-23)
+
+### Enhancements
+- New `*-LabSnippet` cmdlets!
+- New PreInstallationActivity parameter for Add-LabMachineDefinition
+  - Immediately before a role is deployed, the PreInstallationActivity is executed
+  - Machines without roles will get their PreInstallActivity executed after all domains are deployed
+  - For both pre and post, Get-LabInstallationActivity can be used, and can even include custom roles
+  - Thanks to the awesomeness that is @friedrichweinmann and his PSFramework, we now have Variables and Functions that can be passed to Get-LabInstallationActivity
+- Adding SQL Server 2019 to Azure DevOps Server deployment ode and validators.
+- Added product keys for 'Windows 10 Education' and 'Windows 10 Pro Education'.
+- MAC addresses are now configurable
+- A choice selection of WSMAN settings can now be configured
+  - To see which: Get-PSFConfig -FullName AutomatedLab.WinRm*
+- Setting lab administrative accounts to never expire.
+
+### Fixes
+- Fixed #1033, NonExistingDnsServerAssigned is now only called if domain controllers are defined in the lab.
+- Fixed issue with SharePoint servers not being deployed when they had other roles applied to them.
+- Adding alias Disable-LabHostRemoting for increased visibility (Fixes #1137).
+- AL does no longer use the Hyper-V VM's notes field exclusively (Fixes #1135).
+
 ## 5.34.0 (2021-03-30)
 
 ### Enhancements
@@ -22,6 +44,7 @@
 - Using the 20h2 Windows 10 images in Azure now.
 - Fixing upstream issue with duplicate files in SysInternals archive
 - Removing defunct parameter AzureSubscriptionName from New-LabDefinition (#1079)
+- Adding validator for DNS settings on domain members and domain controllers (#3)
 - Fixing issue with post-install scripts and DNS resolution on Linux
 
 ## 5.33.0 (2021-03-03)